requirements.txt
requests
pymongo
python-telegram-bot
python-telegram-bot[job-queue]